requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
docx==0.2.4
frozenlist==1.4.0
gino==1.0.1
greenlet==3.0.1
idna==3.4
lxml==4.9.3
magic-filter==1.0.12
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-docx==1.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.7.1
yarl==1.9.2
